clean up & separate component types
