+++ src/components/text/text.tsx
@@ -2,29 +2,8 @@ import clsx from "clsx";
 import * as React from "react";
 
 import s from "@/components/text/text.module.css";
+import { TagName } from "@/components/text/types";
 import { forwardRef } from "@/components/utilities/react";
-import { Props } from "@/components/utilities/types";
-
-const TagName = "p" satisfies React.ElementType;
-type TagName = typeof TagName;
-
-type TextVariant =
-  | "title"
-  | "subtitle"
-  | "body1"
-  | "body2"
-  | "overline"
-  | "finePrint";
-
-type TextOptions<T extends React.ElementType = TagName> = {
-  as?: T;
-  variant?: TextVariant;
-};
-
-export type TextProps<T extends React.ElementType = TagName> = Props<
-  T,
-  TextOptions<T>
->;
 
 export const Text = forwardRef(function Text(
   { as = TagName, children, className, variant = "body", ...props },
