+++ src/components/theme/types.ts
@@ -1,0 +1,12 @@
+import { BannerIconConfig } from "@/components/banner/types";
+
+export type Theme = {
+  components: {
+    banner: {
+      icons: BannerIconConfig;
+    };
+    spinner: {
+      icon: React.ReactNode;
+    };
+  };
+};
