+++ src/components/theme/use-theme.ts
@@ -1,7 +1,8 @@
 "use client";
 
-import { ThemeContext } from "@/components/theme-provider/theme-provider";
 import * as React from "react";
+
+import { ThemeContext } from "@/components/theme/theme-provider";
 
 export const useTheme = () => {
   const context = React.useContext(ThemeContext);
